ecommerce-intelligence/requirements.txt
scrapy==2.11.0
selenium==4.18.1
webdriver-manager==4.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
streamlit==1.31.1
plotly==5.19.0
pandas==2.2.0
python-dotenv==1.0.1
fake-useragent==1.4.0
APScheduler==3.10.4
requests==2.31.0
colorlog==6.8.2
alembic==1.13.1
scrapy-user-agents==0.1.1
